pyproject.toml
[tool.poetry]
name = "resume"
version = "0.1.0"
description = ""
authors = ["Altair Bueno<business.altair.bueno@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
DateTime = "^4.4"
python-dateutil = "^2.8.2"
chevron = "^0.14.0"
pydantic = "^1.9.1"
PyYAML = "^6.0"
glom = "^22.1.0"
phonenumbers = "^8.12.52"
iso3166 = "^2.1.1"
typer = "^0.6.1"

[tool.poetry.dev-dependencies]
datamodel-code-generator = {extras = ["http"], version = "^0.13.0"}
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
